--- internal_mailbox/error_codes_tap.xml
@@ -1,64 +1,0 @@
-<?xml version="1.0"?>
-
-<internal_error_codes>
-
-
-    <WRITE_PEGA_CMD0>
-        <!-- <rule name="error" code="NO_ERROR">
-            <if>(field.CMD == 0)</if> 
-        </rule>
-        <rule name="error" code="INVALID_COMMAND">
-            <if>(field.CMD == 1)</if>
-        </rule> -->
-        <rule name ="error" code="INVALID_PARAMETERS">
-            <if>(field.RUN_BUSY == 0)</if>
-            <if> (field.CMD != 0) and (field.CMD != 2) </if>
-        </rule>
-        
-    </WRITE_PEGA_CMD0>
-
-    <ASSERT_ERR0>
-        <rule name="error" code="INVALID_COMMAND">
-            <if>(globals.ALWAYS_TRUE == 1)</if>
-        </rule>
-    </ASSERT_ERR0>
-
-    <DFX_SCRATCHPAD>
-        <rule name="error" code="INVALID_COMMAND">
-            <if>(globals.ALWAYS_TRUE == 1)</if>
-        </rule>
-    </DFX_SCRATCHPAD>
-
-    <HWP_CAPABILITIES_MSR>
-        <rule name="error" code="INVALID_COMMAND">
-            <if>(globals.ALWAYS_TRUE == 1)</if>
-        </rule>
-    </HWP_CAPABILITIES_MSR>
-
-    <PSMI_PREP>
-        <rule name="error" code="INVALID_COMMAND">
-            <if>(globals.ALWAYS_TRUE == 1)</if>
-        </rule>
-    </PSMI_PREP>
-
-    <READ_PROCID>
-        <rule name="error" code="INVALID_COMMAND">
-            <if>(globals.ALWAYS_TRUE == 1)</if>
-        </rule>
-    </READ_PROCID> 
-
-
-
-    
-
-    
-    <spec_globals>
-        
-        <ALWAYS_TRUE>
-            <val>1</val>
-        </ALWAYS_TRUE>  
-        
-    </spec_globals>
-</internal_error_codes>
-
-
